Refactor to split generate stdin and stdout.

+++ testcase_maker/generator.py
@@ -23,7 +23,8 @@ class TestcaseGenerator:
     # TODO Custom stdin and stdout file name format.
 
     def __attrs_post_init__(self):
-        self.script_executor = get_executor_for_script(self.answer_script)
+        if not self.script_executor:
+            self.script_executor = get_executor_for_script(self.answer_script)
 
     def new_subtask(self, no_of_testcase: int, name: str = None) -> Subtask:
         if not name:
@@ -33,26 +34,64 @@ class TestcaseGenerator:
         self.subtasks.append(subtask)
         return subtask
 
-    def generate(self):  # TODO Override toggle
+    def generate(self, override: bool = False):
+        self.generate_stdin(override)
+        self.generate_stdout(override)
+
+    def generate_stdin(self, override: bool = False):
         self.output_dir.mkdir(parents=True, exist_ok=True)
 
         for subtask in self.subtasks:
-            for case_no in range(1, subtask.no_of_testcase + 1):
-                print(f"Generating subtask '{subtask.name}', testcase '{case_no}'...")
+            for testcase_no in range(1, subtask.no_of_testcase + 1):
+                print(f"Generating stdin for subtask '{subtask.name}', testcase '{testcase_no}'...")
 
-                stdin_file = self.output_dir.joinpath(f"{subtask.name}-{case_no}.in")
+                stdin_file = self._stdin_path(subtask.name, testcase_no)
+                if stdin_file.is_file() and not override:
+                    print(f"Skipped '{stdin_file}' as file already exist.")
+                    continue
+
                 resolver = Resolver(subtask.override_name_values)
                 stdin = resolver.resolve(self.values)
 
                 with open(stdin_file, "w", newline="\n") as input_buffer:
                     input_buffer.write(stdin)
-                print(f"Saved '{stdin_file}'")
+                print(f"Saved '{stdin_file}'.")
 
-                stdout_file = self.output_dir.joinpath(f"{subtask.name}-{case_no}.out")
-                with tempfile.TemporaryDirectory() as tmpdir:
-                    exec_filename = self.script_executor.compile(tmpdir, self.answer_script)
-                    stdout = self.script_executor.execute(exec_filename, stdin).decode("UTF-8")
+    def generate_stdout(self, override: bool = False):
+        self.output_dir.mkdir(parents=True, exist_ok=True)
 
+        for subtask in self.subtasks:
+            for testcase_no in range(1, subtask.no_of_testcase + 1):
+                print(f"Generating stdout for subtask '{subtask.name}', testcase '{testcase_no}'...")
+
+                stdin_file = self._stdin_path(subtask.name, testcase_no)
+                if not stdin_file.is_file():
+                    print(f"Skipped as stdin '{stdin_file}' does not exist.")
+                    continue
+                with open(stdin_file, "r", newline="\n") as input_buffer:
+                    stdin = input_buffer.read()
+
+                stdout_file = self._stdout_path(subtask.name, testcase_no)
+                if stdout_file.is_file() and not override:
+                    print(f"Skipped '{stdout_file}' as file already exist.")
+                    continue
+
+                stdout = self._execute_script(stdin)
                 with open(stdout_file, "w", newline="\n") as output_buffer:
                     output_buffer.write(stdout)
                 print(f"Saved '{stdout_file}'")
+
+    def validate(self):
+        pass
+
+    def _stdin_path(self, subtask_name: str, testcase_no: int):
+        return self.output_dir.joinpath(f"{subtask_name}-{testcase_no}.in")
+
+    def _stdout_path(self, subtask_name: str, testcase_no: int):
+        return self.output_dir.joinpath(f"{subtask_name}-{testcase_no}.out")
+
+    def _execute_script(self, stdin: str) -> str:
+        with tempfile.TemporaryDirectory() as tmpdir:
+            exec_filename = self.script_executor.compile(tmpdir, self.answer_script)
+            stdout = self.script_executor.execute(exec_filename, stdin).decode("UTF-8")
+        return stdout
